collector/requirements.txt
boto3==1.26.137
pandas==1.5.3
sqlalchemy==2.0.19
python-dotenv==1.0.0
schedule==1.2.0
requests==2.31.0 
numpy==1.23.5
